+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [1.0.23] -
 ### Added
 - ability to inhibit sleep on Windows
+- "Open config directory" main menu option
 
 ## [1.0.22] - 2022-11-07
 ### Added